Added license.
The assets are licensed under the GPL 2+. Note however that this only
applies to the raw source files. A game world created from these files
is NOT considered a "derivate works".
I.e. you can alter these files and populate a server with them without
having to make them available to the players.